pyproject.toml
[project]
name = "datacomp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "boto3>=1.35.99",
    "braceexpand>=0.1.7",
    "cloudpathlib>=0.20.0",
    "hydra-core>=1.3.2",
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.5",
    "numpy>=2.2.1",
    "open-clip-torch==2.16.1",
    "pandas>=2.2.3",
    "pyarrow>=18.1.0",
    "ray[dashboard,default]>=2.42.1",
    "torch>=2.5.1",
    "tqdm>=4.67.1",
    "training",
    "webdataset==0.2.31",
    "zarr>=3.0.0",
]

[tool.uv.sources]
training = { path = "training", editable = true }
